refactor: Phase S - timer unification, examples, base component adoption
- Unify timestamp utilities into platform.hpp (SteadyNowNs/SteadyNowUs)
- TimerScheduler template parameterization + collect-release-execute pattern
- Add SHM IPC demo (producer/consumer/monitor with HSM, Timer, MemPool, Shell)
- Rewrite serial_ota demo with 10 osp components integration
- Adopt base components (FixedString/FixedVector/SteadyNowUs) in upper layers
- Add 6 cross-process fork+shm tests for ShmChannel/ShmRingBuffer
- Add Mermaid diagrams for component deps, process arch, HSM states

+++ include/osp/data_fusion.hpp
@@ -211,7 +211,7 @@ class FusedSubscription {
 
  protected:
   void TryFire() noexcept {
-    for (size_t i = 0; i < kNumTypes; ++i) {
+    for (uint32_t i = 0; i < kNumTypes; ++i) {
       if (!received_[i]) return;
     }
 
@@ -222,13 +222,13 @@ class FusedSubscription {
     fire_count_.fetch_add(1, std::memory_order_relaxed);
 
     // Auto-reset for next cycle
-    for (size_t i = 0; i < kNumTypes; ++i) {
+    for (uint32_t i = 0; i < kNumTypes; ++i) {
       received_[i] = false;
     }
   }
 
   /** @brief Store timestamp from envelope header for the given type index. */
-  void StoreTimestamp(size_t idx,
+  void StoreTimestamp(uint32_t idx,
                       const MessageEnvelope<PayloadVariant>& env) noexcept {
     timestamps_[idx] = env.header.timestamp_us;
   }
